Migrate run_actor_loop orchestrator (1176 LOC) to actor_loop_flow
Move the main actor loop orchestrator `run_actor_loop` out of impl
Agent into actor_loop_flow.rs as a free function taking
agent: &mut Agent. This is the largest single migration of #681,
moving the 1,176-LOC orchestrator that was the last actor-loop
method on impl Agent.

Updates the single production caller `handle_user_message` to call
the new free function `run_actor_loop(self, ...)`.

Visibility upgrades on turn.rs items:

- pub(super) on Agent methods used by the orchestrator:
  - prepare_actor_loop_turn_state
  - current_assistant_model
  - execute_tool_call
  - invoke_photon_evaluate
- pub(super) on free fns / consts:
  - should_record_no_repo_progress
  - build_feedback_for_no_repo_progress
  - build_stats
  - normalize_plan_exploration_key
  - progress_stage_label
  - summarize_plan_write
  - format_blocked_progress_line
  - build_feedback_for_unsafe_block
  - LOG_ARGS_MAX_CHARS
  - PLAN_REPEATED_EXPLORATION_BLOCK_THRESHOLD
- pub(super) on structs + fields:
  - PlanExplorationKey (+ stage / tool_name / normalized_args)
  - PlanWriteSummary (+ action / note / status / phase / signature)

Updates the in-file include_str regression test
`production_repair_dispatch_does_not_call_legacy_decision_bridge`
to read the run_actor_loop body from `actor_loop_flow.rs` instead
of `turn.rs`, and to use the new function-header marker.

Part of #680 / #681 incremental method migration. With this PR the
entire actor-loop orchestration moves out of `impl Agent`; turn.rs
LOC drops from ~37K to ~35.8K and only Agent-state holder concerns
remain alongside small helpers.
